bc global rebase
