+++ libs/blocks/brand-concierge/chat-ui-config.js
@@ -46,6 +46,12 @@ export default {
         tabletLandscape: 1199,
       },
     },
+    chatTranscript: {
+      enabled: true,
+      maxSessions: 10,
+      maxMessagesPerSession: 100,
+      cleanupInterval: 0.5,
+    },
     meetingForm: {
       fieldsPerRow: 2,
       fieldLayoutRules: {
@@ -311,9 +317,9 @@ export default {
     '--chat-container-background': '#FFFFFF',
     '--chat-container-bottom-background': '#FFFFFF',
     '--chat-history-bottom-padding': '0',
-    '--chat-history-padding': '28px',
+    '--chat-history-padding': '24px',
     '--chat-history-padding-top-expanded': '0',
-    '--chat-history-padding-mobile': '24px',
+    '--chat-history-padding-mobile': '34px',
     '--chat-history-padding-tablet': '32px',
     '--color-button-secondary-hover-text': '#FFFFFF',
     '--color-button-submit': '#292929',
