+++ test/blocks/brand-concierge-global/mocks/default.html
@@ -1,0 +1,27 @@
+<header class="global-navigation">
+  <div class="feds-topnav-wrapper">
+    <nav class="feds-topnav" aria-label="Main">
+      <div class="feds-bc-wrapper"></div>
+    </nav>
+  </div>
+</header>
+<main>
+  <div class="section">
+    <div class="brand-concierge-global">
+      <div>
+        <div>
+          <p>Prompt one</p>
+        </div>
+        <div>
+          <p>Prompt two</p>
+        </div>
+        <div>
+          <p>Prompt three</p>
+        </div>
+      </div>
+      <div>
+        <p>Ask Adobe anything</p>
+      </div>
+    </div>
+  </div>
+</main>
